+++ src/components/rightSide/user/UserDropMenu.tsx
@@ -1,9 +1,8 @@
 import { useEffect, useRef, useContext } from "react";
 import { ProfileContext } from "hooks/context/ProfileContext";
 import { onProfile } from "./dropFunction";
-import { getSocket } from "socket/socket";
+import { useAppointAdmin } from "hooks/dropFunc/useAppointAdmin";
 import { RoomIdContext } from "hooks/context/RoomIdContext";
-import { ChatRoomResponse } from "socket/active/chatEventType";
 import * as S from "./style";
 
 export default function UserDropMenu(props: {
@@ -15,7 +14,7 @@ export default function UserDropMenu(props: {
   const setProfileUser = useContext(ProfileContext);
   const roomId = useContext(RoomIdContext);
   const dropRef: React.RefObject<HTMLDivElement> = useRef(null);
-  const socket = getSocket();
+  const { onAppointAdmin } = useAppointAdmin(roomId, props.targetUser);
 
   useEffect(() => {
     const handleClick = (e: MouseEvent) => {
@@ -28,31 +27,6 @@ export default function UserDropMenu(props: {
       window.removeEventListener("mousedown", handleClick);
     };
   }, [dropRef]);
-
-  function resultHandler(res: ChatRoomResponse) {
-    if (res.status === "approved") {
-      if (res.roomId === roomId) {
-        // TEST : 부방장 이벤트 확인용
-        console.log("부방장 지정 승인", res);
-      }
-    } else console.log(res);
-  }
-
-  useEffect(() => {
-    socket.on("appointAdminResult", resultHandler);
-    return () => {
-      socket.off("appointAdminResult", resultHandler);
-    };
-  }, []);
-
-  function onAppointAdmin() {
-    // TEST: 부방장 지정 emit 폼 확인
-    console.log("부방장지정 폼", roomId, props.targetUser);
-    socket.emit("appointAdmin", {
-      roomId,
-      username: props.targetUser,
-    });
-  }
 
   return (
     <>
