+++ src/hooks/context/RoomIdContext.ts
@@ -1,3 +1,3 @@
 import { createContext } from "react";
 
-export const RoomIdContext = createContext<number | null>(null);
+export const RoomIdContext = createContext<number>(0);
